+++ koku-test.yml
@@ -7,14 +7,14 @@ services:
       working_dir: /koku
       entrypoint: 'tox -r'
       environment:
-        - POSTGRES_SQL_SERVICE_HOST=db
+        - POSTGRES_SQL_SERVICE_HOST=test_db
         - POSTGRES_SQL_SERVICE_PORT=5432
       volumes:
         - '.:/koku/'
       links:
-        - db
+        - test_db
 
-  db:
+  test_db:
     image: postgres:9.6
     environment:
     - POSTGRES_DB=koku_test
