+++ koku/api/tags/ocp/ocp_tag_query_handler.py
@@ -68,15 +68,14 @@ class OCPTagQueryHandler(TagQueryHandler):
 
         return tag_keys
 
-    @staticmethod
-    def get_dictionary_for_key(merged_data, key):
-        for di in merged_data:
-            if key in di.keys():
-                return di
-        return None
-
     def get_tags(self, tenant):
         """Get a list of tag key and values to validate filters."""
+        def get_dictionary_for_key(merged_data, key):
+            for di in merged_data:
+                if key in di.get('key'):
+                    return di
+            return None
+
         with tenant_context(tenant):
             tag_keys = OCPUsageLineItemDailySummary.objects\
                 .filter(self.query_filter)\
@@ -87,14 +86,16 @@ class OCPTagQueryHandler(TagQueryHandler):
             merged_data = []
             for item in tag_keys:
                 for key, value in item.items():
-                    key_dict = self.get_dictionary_for_key(merged_data, key)
+                    key_dict = get_dictionary_for_key(merged_data, key)
                     if not key_dict:
                         new_dict = {}
-                        new_dict[key] = [value]
+                        new_dict['key'] = key
+                        new_dict['values'] = [value]
                         merged_data.append(new_dict)
                     else:
-                        if value not in key_dict[key]:
-                            key_dict[key].append(value)
+                        if value not in key_dict.get('values'):
+                            key_dict['values'].append(value)
+                            key_dict['values'].sort()
         return merged_data
 
     def execute_query(self):
@@ -104,12 +105,12 @@ class OCPTagQueryHandler(TagQueryHandler):
             (Dict): Dictionary response of query params and data
 
         """
-        if self.query_parameters.get('filter'):
-            tags = self.get_tags(self.tenant)
-            query_data = tags
-        else:
+        if self.query_parameters.get('key_only'):
             tag_keys = self.get_tag_keys()
             query_data = sorted(tag_keys, reverse=self.order_direction == 'desc')
+        else:
+            tags = self.get_tags(self.tenant)
+            query_data = tags
 
         self.query_data = query_data
         return self._format_query_response()
