+++ koku/api/tags/queries.py
@@ -17,8 +17,8 @@
 """Query Handling for Tags."""
 import logging
 
-from api.query_handler import QueryHandler
 from api.query_filter import QueryFilter
+from api.query_handler import QueryHandler
 
 LOG = logging.getLogger(__name__)
 
@@ -52,10 +52,10 @@ class TagQueryHandler(QueryHandler):
         """
         filters = super()._get_filter(delta)
 
-        project = self.get_query_param_data('filter', 'project', list())
-        if project:
-            proj_filter = {'field': 'namespace', 'operation': 'icontains'}
+        project = self.get_query_param_data('filter', 'project')
 
+        if project and not TagQueryHandler.has_wildcard(project):
+            proj_filter = {'field': 'namespace', 'operation': 'icontains'}
             q_filter = QueryFilter(parameter=project[0], **proj_filter)
             filters.add(q_filter)
